[Fix] add aprove contract
